src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt src/requirements/requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via loguru
dj-database-url==2.1.0
    # via -r src/requirements/requirements.in
django==5.0.3
    # via
    #   -r src/requirements/requirements.in
    #   dj-database-url
    #   django-ckeditor
    #   django-cors-headers
    #   django-extensions
    #   django-js-asset
    #   django-rosetta
    #   djangorestframework
django-ckeditor==6.5.1
    # via -r src/requirements/requirements.in
django-cors-headers==4.3.1
    # via -r src/requirements/requirements.in
django-extensions==3.2.3
    # via -r src/requirements/requirements.in
django-js-asset==2.2.0
    # via
    #   django-ckeditor
    #   django-mptt
django-mptt==0.16.0
    # via -r src/requirements/requirements.in
django-rosetta==0.10.0
    # via -r src/requirements/requirements.in
django-split-settings==1.3.0
    # via -r src/requirements/requirements.in
django-widget-tweaks==1.5.0
    # via -r src/requirements/requirements.in
djangorestframework==3.15.1
    # via -r src/requirements/requirements.in
gunicorn==21.2.0
    # via -r src/requirements/requirements.in
idna==3.6
    # via requests
loguru==0.7.2
    # via -r src/requirements/requirements.in
packaging==24.0
    # via gunicorn
pickles==0.1.1
    # via -r src/requirements/requirements.in
pillow==10.2.0
    # via -r src/requirements/requirements.in
polib==1.2.0
    # via django-rosetta
psycopg==3.1.18
    # via -r src/requirements/requirements.in
python-decouple==3.8
    # via -r src/requirements/requirements.in
python-dotenv==1.0.1
    # via -r src/requirements/requirements.in
python-gettext==5.0
    # via -r src/requirements/requirements.in
requests==2.31.0
    # via django-rosetta
six==1.16.0
    # via -r src/requirements/requirements.in
sqlparse==0.4.4
    # via django
typing-extensions==4.10.0
    # via
    #   dj-database-url
    #   psycopg
tzdata==2024.1
    # via
    #   django
    #   psycopg
urllib3==2.2.1
    # via requests
whitenoise==6.6.0
    # via -r src/requirements/requirements.in
win32-setctime==1.1.0
    # via loguru
